Add flexibility / configurability to the hex text displays
The `HexText` and `WalletAddressText` component both truncate
to the first 4 and last 4 characters in the rendered string of the
hex value.  These first 4 characters effectively mean only the first
two leading hex nibbles are visible, which may not be unique
enough to easily distinguish between other values.

@QuentinI suggested that the leading 4 nibbles should be visible
at a minimum.

Additionally, when inspecting etherscan, it can be seen that such
leading and trailing truncation can be larger, or smaller depending
on the context.

With these criteria in mind, it becomes advantageous to be able
to specify the truncation amount for the leading and trailing
characters of the hex display text values.  With careful
consideration, and automatically accounting for the leading `0x`
this should allows these components to be as flexible as needed.