Add some default Node Validator Numbers
The number of available sequencer nodes isn't something that we
are able to easily provide for the Block Explorer UI at the moment.
Doing so requires us to change the Explore APIs, which we are not
wanting to do so soon before the Proof of Stake release.  Instead
the values are to be hard-coded depending on the environment
until they can be populated more appropriately.