Improve search input support
The SearchInput element is fairly restrictive in what it supports for
detecting block numbers in that it only accepts just a string of digits
without any human formatting.  Additionally, it is not strict when it
comes to submitting search queries to the Explorer service to look
for results.

For the Explorer search function, we know that it only accepts valid
TaggedBase64 representations.  As such, we should be able to
prevent queries that we know will fail as they are not valid
TaggedBase64 representations.

For the block number parsing, it doesn't accept common human
numeric formatting conventions.  While it may not be feasible to
support the actual human formatting, we can at the very least strip
out well known locale grouping separators before attempting the
number detection in order to support them.

Additionally, it was noticed that the "Enter" button on a NumPad
does no trigger a search, so that has been added as well.