Update Explorer Home Page to do polling updates
The Explorer Page is causing some confusion for some people with
the information it shows not being updated continually.  Because
the information doesn't update ever, and some of the components
end up being dynamic, specifically with timers counting down, it
makes it look like the Chain is not progressing.

To rectify this issue without doing a lot of reworking of the page to
support the Node Validator API, I've introduced a polling solution to
re-poll the Explorer Summary every 2 seconds for the moment.  It
can be revisited in the future to add a proper real-time component.

Right now this solution won't pick up updates as soon as they are
available to be consumed.  Instead, it must wait the full polling
interval before it has a chance to know for certain.

Additionally, it occurs to me based on the feedback I've heard that
the "Time" field in the Latest Block display may be being interpreted
as the time to produce the block, which it is not.  To help clear up
this confusion, I have renamed the "Time" field to "Age" as it
reflects how old the block is.  I have also added a "Time" field
back in that shows the time taken to produce the block.