Rename "Time" field to "Block time"
The "Time" field represents the amount of time it took for the block
to be produced.  It has been requested to be renamed to "Block
time" in order to alleviate any confusion.  It has also been requested
to prefix it with a `~` to indicate that it is an approximate value.