Update Nodes Summary Data Table
The Nodes Summary Data Table supports showing the address and
stake of the nodes.  This was added for the Delegation UI initially
but then it was decided to move it elsewhere.  It's still
useful to support these fields when they exist.  However, when they
don't, then the data just shows as missing.  As a result, it would be
beneficial to only show the data when it's available.

The easiest determination is to check the total stake available, since
we must calculate it anyway.

This change adds two separate table layouts, one with stake table
information, and one without.