Fix the Explorer Page not loading in the Block Explorer
When making the change to the Explorer Page to perform its
request by periodic polling instead of just polling once and then
being done, the NextJS client component was missed with the
change of how the data was provided.  This makes it never pull
any data, and never load.

This change replaces the component needed to make this work.