Fix textAlign props in SVG Text
The SVG Text element has a `textAlign` property that we are
currently assigning the value `"center"` to.  This is not a supported
value.  We could replace it with the appropriate corresponding value
of `"middle"`.  But, the text is actually aligned with the start of the
element.  As such, it is better to replace them with `"start"`.