Add `opacity` to will-change list for search input barrier
There seems to be a rendering issue utilizing Chrome 141.0.7390.108 (Official Build) (arm64).

When the input barrier on the search bar is focused or loses focus, it's meant to transition
between a transparent color background to a full color background.  However, this backdrop
has an opacity applied that is constant. With this combination there are brief flickers of
the full color without any opacity applied.

Adding "opacity" to the will-change list in the css seems to help fix this issue.