Add Delegation UI Site Section
In order to start working on the Delegation UI, and to separate it
from the Block Explorer while still being able to reuse the prebuilt
and tested components for the Block Explorer, we require some
level of separation.  Normally, this would facilitate a complete
restructure of the packages into smaller packages for ease of
consumption.  That may be a goal in the future, but to get things
running for now, we opt to give the Delegation UI it's own section
within the Block Explorer Components.