Add a Fake Data implementation of the L1 Validator Service
In order to be able to effectively test the Delegation UI's layout,
design, and interaction with the external L1 Validator Service, it
is necessary to be able to communicate with the service, even if
the service hasn't been completed yet and is not up and running.
To help facilitate this, Mock data should be utilized in order to
provide an approximation of what the resulting service
communication will return.