Add WebWorker implementation of L1 Validator Service
It is beneficial to run external requests to APIs on a separate thread from the
main thread so that these requests have a minimal impact, if any, on the UI
rendering and page performance.

This commit adds client and server implementations of the L1 Validator Service
so that this communication can be spun up as seamlessly as possible.