Add a ReactContext for the ConsensusSet and a Derive component
The Consensus Set is is needed in order to quickly lookup data for all of the
validators in order to determine if they are part of the active consensus set or
not.  Using a `Set` helps to quickly determine whether or not something is
a member or not.   However, the `address` is currently formatted as an
`ArrayBuffer` instead of a `string` in the type implementations, and
`ArrayBuffers` are not directly comparable using the `Set` implementation.
As a result, this `Set` will need to store strings.