Add a React Context for Validator Ranks and a Derivation of them
In order to display the Rank value of different Node Validators compared to each
other, it helps to have a definitive source of the data.  The React Context can
be utilized to Supply the Ranks as a Map of addresses formatted as strings to
their rank value.

Additionally, a `DeriveRank` component has been added to rank Validators
against each other utilizing the validators from the `AllValidatorContext`, and
a pre-defined `score` function to score them.