Add React Context for Total Stake and it's derivation
In order to compare Validator's stakes against each other, and to display
information such as the Total Stake currently delegated, it is beneficial to have
a way to provide this value to the component Tree.  The Total Stake Context
does this easily.

Additionally a `DeriveTotalStake` component has been added to derive the
TotalStake from the set of `AllValidators`.