Add TotalSupply React Context, and a way to retrieve it
The TotalSupply of the ESPToken can be sourced from the ESPToken contract.
This information can be retrieved and made available to all components via a
React Context.

In order to help specify it a helpful component has been created and added to
facilitate the ability to read arbitrary data from contracts, and to provide the
async data retrieval as an `AsyncSnapshot` much like `PromiseResolver` or
`AsyncIterableResolver`.  It is called `ReadContractToAsyncSnapshot`, and
it has props similar to the `useReadContract` function provided by `wagmi`
which it utilizes internally.

Utilizing the `ReadContractToAsyncSnapshot`, The `ProvideTotalSupply`
is able to retrieve the Total supply and provide it to the `TotalSupplyContext`
that has been created for it.  Additionally, and for convenience, the inspection
of the AsyncSnapshot can be performed utilizing the
`TotalSupplyAsyncSnapshotContext` context.