Add Connect Wallet Button Component
The Connect Wallet Button Component is required in order to connect the wallet.
For convenience, we aim to connect the RainbowKit library as a multi wallet sign
on tool.

The Connect Wallet Button has two separate representations, one for when the
Wallet is connected, and one for when it is not.