Add DelegationUIContent component
The Delegation UI Content represents a wrapping container for the content of the
Delegation UI itself.  Since the Design indicates that the contents of the page
should be centered, and not the same size as the rest of the Page (or header)
We require a dedicated content element to contain and restrict the size of this
content.

Additionally, the design utilizes a dedicated Block that is blurred for stylistic
purposes.