Add Validator Table Sorting Component
In order for the contents of the Validator Table to be sortable, we need a
consistent way of transforming the data to sort it, so that we can maintain
the original state, and provide the new sorted state as needed.  This sort is
meant to be done in a non-destructive way.