Add Cells and Header Cells for Validator Table
In order to effectively display the contents of the Table, regardless of the ultimate
form of the table, whether using an HTML table element, or a grid css style, the
need for contents to bbe displayed need to be considered.

These components represents the different cell types, both content and heading
that are needed for the table to be created.