Add Delegation Table Component Elements
The Delegation Table has been added in order to display the contents of the
table in the Design, and to support the ability to click the header elements in
order to change their sorting direction / sort type.

This table has been created as an HTML table for it's accessibility implications.