Align type definitions with the rust code
In order to be able to seamlessly communicate with the Backend service, it is
important that the type definitions are able to be decoded and encoded in the
correct format.  Additionally, in order to try and ensure ease of maintenance, it
is also helpful to try and make the types have the same names where possible.

These changes realign the type definitions with their implementations on the
rust side.