Add SegmentedButton component
The designs call for a button to switch between all validators and my staked
validators.  This is currently designed as a Segmented Button, which is a button
group that adds like a radio selection with a group of potential values, and only
one is ever selected.

This component is designed to be flexible to support multiple values.