Add Contract Mocks and RainbowKit Mocks for Fake Data
In order to be able to effectively control the state so we can test edge-cases, we
need to be able to Mock the interaction with the Chain.  It's not very feasible to
intercept these calls and simulate them, as they already undergo ABI encoding
transformations, and may be submitted externally through the wallet connection.

Additionally, utilizing Wagmi's React hooks make it very difficult to intercept these
calls in a meaningful way.  As a result, it becomes better to avoid their
convenience in the favor of being able to substitute abstractions.  As a result, we
still utilize Wagmi's actions, but we create specific contract interfaces and
implementations, to be utilized by the UI where possible.  This makes the Web3
calls behave more like any normal abstraction.  While we lose the convenience of
the Async handling provided by Wagmi, and tanstack, we utilize our own already
created Async Handling to bridge this gap.  As a result, the resulting calling code
looks clearer, and is much more readable.

These mocks automatically get injected when the environment is "FakeData" and
no contract addresses have been specified.