Fix nodes/all API definition
The documentation for the nodesAll path has been updated to include a hash.  As
a result, there's now no convenient way to retrieve the latest snapshot in a single
request.  This doesn't make much of a difference, but it does mean needing to
retrieve the L1 Block information before proceeding.