Add new Staking Modal with New Delegation
The Delegation UI calls for interactivity via some new Modals.  These modals are
the primary way that users will interact with and submit delegations and
undelegations.  Their resulting states are somewhat complex and have a lot of
intricacies about them.

This represents the initial setup and configuration for data needed for the
Delegation UI Modals, as well as the initial implementation of the New Delegation
Modal.

Mocks have been adjusted to no longer be immutable so they do not immediately
trigger a re-render (which reflects more accurately what the live environment will
be), which also simplifies many of the testing checks.  Immutability is a desirable
trait, but sadly, it would cause us to overlook refresh interactions as they would
happen automatically.