Add Undelegation Content for Staking Modal
We need to be able to undelegate our existing delegation to the contract.  In
order to achieve this, we also need to be able to recognize that we have an
existing stake, and to allow for the choice to perform further delegation, or
to undelegate.  These decisions also need to have some automation.  If the
user does not currently have an active stake against a specific validator, then
it should automatically choose the new delegation path without the user's
interaction.

This PR achieves an initial implementation of these goals.