Add Injection for Mocks to handle L1 Validator Service calls for Wallet
The FakeData version of the L1 Validator Service runs in a WebWorker, and as
a result, it does not have access to the memory or the state of the main thread.
With the current contract Mocks there's no real easy way for the FakeData to
access the data needed.

The proposed solution is to take advantage of the indirection and implement a
a Mock Wallet API that can handle the API requests, while others are still
forwarded to the Web Worker implementation.  This allows us to intercept and
inspect state changes locally so our Wallet API can return data based on the
results we've seen.

In addition, an Injector to automatically hook up this Mock API and supply it
to downstream components has been added.

Mock contracts and L1Methods have also been made to be mutable when
receiving a new account address, otherwise this causes full page refreshes
as the object is a new instance.  This is for convenience and should help the
page's behavior to match what the live page would see.