Add ability to perform withdrawals
When a wallet has undelegated, or a validator has exited, that wallet needs to be
able to reclaim their freed funds after they have passed the escrow period.

The design calls for Calls to Action to inform the user of the state of these
undelegations and exits in order to prompt the user to perform the appropriate
withdrawal action where applicable.

This also calls for some new modals and tables with slight variations on the
existing implementations.