Fix unit tests due to Stake Table change
When interacting with the StakeTable contract interface and React Contexts, it
was noticed that React doesn't really enjoy working with arrays containing
`BigInt`s.  It attempts to serialize them, presumably for memoization checking.
As `BigInt`s themselves are unable to be serialized directly using `toJSON` or
`JSON.stringify`, this causes an error which breaks React entirely.  To avoid this,
and for some convenience, the `StakeTable` interface has been modified to
return `class` instances where these tuples were previously being utilized.

This makes the interaction between the abstraction of the contract and the
actual contract a little more obscured for the sake of being representable and the
freedom to pass these objects around in React Contexts as we wish.

When this change was made, the corresponding Unit Tests were not updated
causing them to no longer pass.  This change modifies the tests to reflect the
new changes in the types being returned.