Add Claim Rewards functionality
The last major feature needing to be added is the ability to claim the earned
rewards for the user.  This feature is somewhat interesting as the total amount
of claimable rewards isn't given freely but is rather derived from two separate
values.  The total amount of rewards claimed over the wallet's lifetime, which
is retrieved from the RewardClaim contract, and the lifetime rewards which
is retrievable from the reward-state-v2 of the hotshot query service.

Using these two pieces of data, we are able to derive what the user's claimable
rewards will be.

We need to be able to claim these rewards, and refresh our state just like we do
with all other actions in this Delegation UI.