Fix units test for reward state
With the need for the Reward Claim contrat to now need to dispatch funds
using the ESPToken contract, the constructor arguments have changed.  This
broke the previously working tests where this wasn't the case.  In
addition, the Reward Claim contract is really meant to mint new tokens
as a reward.  But these mocks don't have the ability to do this quite yet,
so as a work-around, I just prefund the Reward Claim account with tokens to
hand out instead.