Update type definitions to align with rust
Two types have been modified from the last version represented here.  Those
types are `ParticipationChange` which has been modified to use node index
offset instead of the address of the node, and `Delegation` which as been
modified to remove the effective block and epoch information from it, as
this would be an estimate anyway.