Replace hard-coded colors with variables
In the Delegation UI, due to rapid iteration many colors were hard-coded out of
convenience.  We aim to only reference colors by their variables names so that
we can easily catalog all of the different colors in use in our pages.

This change replaces all current hard-coded colors in the Delegation UI with
predefined colors.