Add Staking Modal Stack History
The designs call for the staking modal to have a "Back" button that is available
at times that seems to be aware of your navigational journey.  By hitting "back",
we expect the user to be taken to a previous modal page in his/her journey.

Since we have the ValidatorSelectionEnum already, we can just store a stack of
previous selections in order to determine the correct navigation to take.  As a
result, the BackButton can become a permanent fixture of the StakingHeader
component, since it can be conditionally rendered.

Furthermore, this requires us to be consistent with our usage of the Modal,
the selection state, and the history itself.  As a result, for convenience we
also added a Close context to pass through a function that will close the
modal, and perform the needed cleanup for the different states.