Fix encoding of Hashes and Addresses for Delegation UI
The types in the delegation ui are communicated and transferred via JSON
encoded objects.  The assumption was made that the data that is ultimately
going to end up being in `ArrayBuffer`s were encoded via Base64 Standard
encoding.  However, they are actually encoded as hex strings.

This change corrects the incorrect encoding usage.