Add more filters for node validators
Add Switch component

The Switch component represents a standard input toggle of "on" or "off".  It is
used as a nice alternative to a standard input checkbox.

Add filter by top 100

The validator list can be filtered by the top 100 nodes, by stake.  This allows us to
preview the expected upcoming Stake Table Active Validator Set.

Add search term filter

The validator list can be searched using a search box.  The design calls for the
node list to have display names.  It is intended to be able to search for nodes by
name, but since we do not currently have those, we're searching by their address
instead for now.