Update the exit escrow period on the Mock Stake Table contract
The Mock Stake Table Contract has a current exit escrow period set to 2000
milliseconds.  This is far too quick for the count time timer to be previewed before
we close the modal.  We update it to be about a minute instead so we can see the
result before it occurs.