Adds multiple adjustments and features
Add estimateGasFee to L1Methods

In order to calculate the fees for a specific contract execution we can utilize the estimate gas features in order to arrive at an appropriate gas estimation.  The fees themselves can then be displayed in ETH for the transaction fee display.

Adds GasEstimator for all contracts

Since ethereum can effectively call the same contract in a "write" mode as well
as an "estimateGas" mode, we end up with a slight failure of our abstraction.  The
abstraction can only ever write or read data to or from the contract, without being
able to inspect other aspects of it.  To resolve this, we recognize that estimating
the gas cost of a contract is similar to calling the contract, but the return type
is different.  As a result, we can just taking the writable / mutable functions of
the state, and have the same calling arguments, but a different result type.  Thus
making another abstraction specifically for gas cost evaluation.

Add Fallback URL Encoding for Storybook Story URLs

If the user wants to overwrite the URL being utilized is a particular storybook
story  he/she is already able to do that.  However, the URL then becomes
unable to be shared with the new parameter.  This is due to how storybook
encodes its value overrides in the URL.  This mechanism is in-band and fails
entirely when special characters are encountered.  As a result, we have opted
to allow for a fallback URL that is encoded as either a hexstring, or a raw base64
string.  This will allow the URL to be shareable once more at the cost of initially
encoding the desired URL.

Add rewardClaimContractAddress to storybook controls

The RewardClaim contract is already supported, but there's currently no way for
a user to specify the contract address in the Storybook stories, and as a result
no way for the Story itself to specify the contract address.  This change adds
the ability for the Storybook story to do just that.  However, the RewardClaim
contract addresses are not currently known.

Update Contract Addresses in Local Dev Net

It seems that the contract addresses in the local dev net have changed once
again.  These addresses have been updated with the current understanding of
their value for convenience.

Add Default L1 Validator Service URL for Delegation UI Local Dev Net

The L1 Validator Service, when running in a local dev net, is currently hosted on
the system on port `8080`.  As a result, we have coded this default URL in for
convenience of development.

Add comments

Multiple comments have been added across many files in order to help explain the
point and purpose of specific React Contexts, Components, and functions.

Add Staking Modal States Storybook Stories

In order to better track the Staking Modal States, and illustrate the different
states that the modal can be in, Storybook Stories have been added so that they
can be directly linked to as a demonstration.

Add some population of the Tx Fee field

The Tx Fee, based on the design, is meant to show the approximate transaction
cost in USD.  This seems somewhat infeasible, and as a result, I've repurposed
it to show the cost in terms of estimated gas, if we are able to determine this
value.