Adjust Perform transaction calls
In order to ensure we're covering all sides of our processing flow, it is beneficial
to be consistent with the logic for how we determine whether our submit
transaction process is in the processing states that we expect.  As a result, the
logic for all of the buttons.

Additionally, a root cause for some Storybook errors was discovered and it
seemed to be due to the `AsyncSnapshot` itself being serialized while containing
a `bigint`.  As a result, the `toJSON` method of `AsyncSnapshot` has been
overwritten to attempt not to serialize the data being stored.