Update benchmarks
It was noticed that the benchmarks for the Map functions were not correct, as it
was running the test for size `n` for `n` total times.  This made the benchmarks
seem worse.  This has been adjusted.

Additionally, the other benchmarks for functional implementations have been
added.

It was also noticed that running the benchmarks in Parallel with `Promise.all`
ends up with them effectively influencing each other in subtle ways.  To eliminate
this, the benchmarks are now run in isolation to lessen the impact of affecting
each other.  This means they take longer to run, but should provide a more
accurate picture rather than favoring the last running benchmark due to less
contention.