Add some UI / UX testing
In order to effectively test the implementation, it is beneficial to add some
automated testing to the implementation.  This marks an initial effort to
add some testing to the UI / UX state, by performing a new delegation.

Add account spoofing to Storybook Story

@jbearer has requested the ability to spoof an account for convenience in
testing.  This eliminates the need for the user to connect his / her wallet, and
they should be able to inspect the current state of his / her wallet as a result.
This information is public, and this should be fine.  But beware, the user won't
be able to perform any mutations, as his/her account is not actually
connected.