Add support for mobile layout
The designs for the mobile layout add some complexity in some places that
require element changes instead of just content changes.  We aim to keep the
DOM changes to a minimum in order to support the mobile layout wherever
possible.  However, where this has proven to not be possible, new elements
have been added that will be hidden on the desktop view, but are visible in
mobile layouts.

These changes have necessitated many changes in css to allow for both
layouts to work within a single css file.  We aim to keep the spirit and intention
of the design as the priority.

Add copy wallet address convenience

@jbearer has requested the ability to copy the validator address on this page.  In
order to avoid tainting the components in the delegation ui with those in the
block explorer, we have duplicated some existing components, and used new
icons that apply directly to the design itself.

Add sort indications on the table column headers

When a column is currently sorted in the UI, it is difficult to determine which
column is the one that has the sort applied to it.  To help, we've replaced the
sort icon with a chevron that helps to indicate the sort direction.

Fix processing state for staking

When using the staking button, the progress indicator would never show that the
delegation had completed.  This was due to the staking amount being reset to
0 when completed.  But even if we didn't set it directly, the user could still
modify it and confuse themselves as a result.