Fix test broken by changes to progress indicator
A unit test for performing a new delegation was broken when the progress
indicator status was updated to present a unique component for an approval
being successful.  The design doesn't account for this state to indicate any
sort of success, but rather prompts the user to continue their journey.

We remove this unique state from the UI in order to fix the test.