Fix progress indicator area not displaying
The progress indicator area is powered by an `image-mask`.  It was noticed that
loading the `svg` from a reference in the `dom` using its id actually prevents that
mask from being able to be scaled appropriately.  To fix this issue, the svg had
been saved as an asset, and the url was adjusted accordingly as this resolves the
mask properties from not being considered.

When we added this fix, we utilized a relative reference, which seems to break it
as it's not a root reference.  To fix the issue we need to change it to a root
reference.