Add Delegation UI Polling updates
We would like for the Delegation UI to update automatically.  However, we have to
juggle some considerations for these interval updates.  In general, we want to
check for updates at a regular interval of about 1s.  Additionally, we would like for
the updates to be responsive to events such as the user connecting and
disconnecting his/her wallet.  We, also need to make sure that the updates do
not trigger too often, or cause any hot looping.

The easiest solution is to consider the `now` timestamp provided by the `Now`
context, the `activeAccount`, and whether we have some data readily available
not.

This also requires a lot of updates with the injection for the Fake Data and Mock
contracts in order to be able to return valid update data.

Fix Component re-render issues

These changes have some sweeping effects that need to be adjusted as needed.
Additionally, changes to this style of update have revealed some Component
re-render issues.  These issues are due to how some of these components have
been created, and what triggers their refresh.  They are ultimately due to the
components being created directly, and not wrapping `children` components
in some cases.  The fix has been to separate the Context Providers in a separate
component from the content itself.

Refactor Hot Shot Query Service Implementations

The different HotShotQueryService implementations were not consistently
represented like the l1 validator service.  This is primarily due to the Hot Shot
Query Service implementations being defined and developed first.  After creating
the L1 Validator Service, and having it's implementations nicely separated, it was
considered desirable to have a similar layout for the Hot Shot Query Service
implementations.  As a result, the implementations have been moved to different
folders as needed.

Add more currency formats for full and normal

The current currency formats for ESP and ETH have 18 significant digits of
representation.  This makes much of the rendered currencies a lot longer than
they absolutely need to be for the sake of precision.  However, it does still
make sense for there to be longer form of currency representations when
viewing details.  As a result, there is a need to have a "summary" or "short"
format for these currencies, and a "full" or "long" format.

This necessitates creating more currency formats for crypto currencies, and
the corresponding new `MoneyTextFull` component.

Add "next" prop to AsyncIterableResolver

The javascript Iterators and AsyncIterators have the ability for the user to pass
values to the Iterator in order to affect or modify the logic / context of any
Generator function.  This has proven to be valuable in the case of performing
incremental updates, with context update changes.  This was especially needed
for triggering some complex interactions.

Add a Debugger Executor to create programmatic breakpoints

In order to ensure that we can use `debugger` statements in Javascript without
having to worry about them being available in the execution environment, we
take the same approach that we did with the AssertExecutor.  If we're defined
within a "dev" environment, then it will be enabled, otherwise it will not.

Various other cleanups and changes