Revise delegation ui tests
The tests for the new delegation interaction in the delegation ui are not currently
passing.  This caused some inspection of the current tests, and the realization
that we can structure this criteria a bit better in order to avoid such specific
css selectors, and make the test less brittle to css changes.