Fix css issue
There is a css issue that was introduced when attempting to fix the
"Nothing brewing yet.." content from displaying.  This has to do with the layout
shifts for mobile versus desktop view.  We still want to be able to display the
"Nothing brewing yet..." content on mobile, but we have some edge cases
that need to be considered and targeted.

This change aims to fix the content from not displaying, and not displaying within
the content space available.