Replace Collapsable Section element with html appropriate elements
The current implementation of the Collapsable Sections utilize `div`s with data
attributes to achieve the desired effects.  There are aria roles to consider for
collapsable sections in order to make them more accessible.  However, there
are **also** specific `HTML` elements that exist for these interactions.  Those
elements are the `details` and `summary` elements.

This change replaces the `div`s and the interactions for expand / collapse with
the appropriate `HTML` elements in order to resolve the accessibility
requirements.