Adjust Delegation UI Local States
The logic within the single component for managing the polling for new updates
for the L1 State and Espresso changes for the L1 Validator Service hasn't been
performing as well as I'd like.  The issue stems from trying to be extremely
flexible, and synchronizing the refresh windows.  By doing this, we end up in
a situation where nothing is immediately reactive, and there's no easy way
to resolve it with the approach.

The solution is to revamp the approach and take it one element at a time for
retrieving the data needed for updates.  By keeping the components separate,
they are able to update and react to updates independently of each of their
respective responsibilities when it suits them.

The AsyncStreamers are actively polling all of the time, and as a result the
criteria for when to trigger a refresh needs to be considered by the Iterator
that is going to be producing the next result.  In this way, we can slow down
a minimum amount while allowing for the page to still be near immediately
reactive to requests / changes where possible.