Adjust block-explorer NextJS build
When originally looking into adding more environment variables to the
block-explorer to support different configurations, it seemed difficult to actually
be able to load environment variables at runtime of the service.  This caused the
need to create a quick solution as environment variables were baked into the
bundle at build time.

However, this has a lot of detriments, it becomes inflexible, requires more tools
to actually build, and worst of all, it requires a build step when launching the
program.  This means that initial program launch can take seconds to minutes
and is terrible for easy deployment / testing.

After investigating how to actually utilize runtime provided environment variables,
some different approaches became available, and we reworked the components
to read from the environment variable.  This allows us to do the next built step
when we're actually building the docker image while allowing for dynamic
configuration with environment variables.  We were even able to sort how how to
generate the `config.json` file dynamically.  As a result, we are back to having
a much quicker to deploy docker image, and a better deployment experience.