Add block display to epoch ends in
For convenience, it is beneficial to know and understand the
current block progression of the Espresso chain.  While the
Delegation UI does not explicitly call for the display of the current
Espresso Block, we still have access to that information, and it is
helpful to make it available.

To that end, this change aims to make it available by hovering on
the current epoch displayed in the Epoch Ends In component.