Add some React.useEffect callback cleanup
In order to make sure that components do not continue to update
after they are unmounted (or to minimize the effect at the very
least), it is important to ensure that callback functions that are
running in asynchronous processes are not invoked where
possible.

This change aims to achieve this by setting the state changing
callbacks to a noop function when the component unmounts.