Add a log and a sleep to fetchNextL1Block if error is unknown
When the Staking UI Service was restarting yesterday, it was
noticed that an error in retrieving specific block id information
from the Server while it was starting with a "Not Available" error
caused the UI to flash with content, and then disappear.

Until we can detect this case more specifically, we address
the error by logging it for inspection and continuing as if it
were a retryable error.