Adjust `TextEditing` component behavior
The `TextEditing` component is very aggressive when detecting
and reporting updates.  This causes the component to re-render
more often than necessary. Furthermore, when the component's
state is being modified and is reported by the `onChange` and
`onSelect` callbacks, the actual state of the component is not
being modified.  This leads to odd behavior issues as further
adjustments to the component result in more modifications
against an unmodified state.

Additionally, when the incoming value is modified, we always
perform a selection when given a valid selection, regardless of
whether the current selection is already applied.

All of these modification issues / detections are somewhat
minor yet detrimental when we attempt to intercept edit events
from the `TextEditing` component and provide formatted
values.