Update MoneyText components to new method
The `MoneyText`, `MoneyTextFull`, and their corresponding
specific monetary formats all utilize floating point representation
when displaying monetary values.  This means that we get
loss of precision pretty quickly when dealing with values like
`ESP` and `ETH`.

Since the `Intl.NumberFormat` supports `StringNumericLiteral`s,
and the `MonetaryValue` now has a convenience method for
outputting `StringNumericLiterals`, we opt to migrate to this
in order to avoid as much precision loss as possible.