Update formatting behavior for ESPInput
The ESPInput never attempted to format the value for the user
before, instead it opted to try and make sense out of what was
being given to the user.

This approach has led to some feedback from @jillrcarlson,
specifcally regarding leaving a leading `0` when the user
chooses to just type at the end of an input, instead of selecting
or deleting the previous value.

The implication behind this dissatisfaction seems like it might
extend to any formatting the user specifies that may be
incorrect.  As a result, the choice to try and auto format the
user's input as he/she modifies the input.  This has led to a lot
of strife and complications in the logic.  There are likely
undiscovered edge cases that still exist.  However, much of
the modifications have helped to improve the underlying
`TextEditing` component.